Use akvelon hubspot plugin dependency

+++ sdks/java/io/cdap/src/main/java/org/apache/beam/sdk/io/cdap/MappingUtils.java
@@ -83,8 +83,8 @@ public class MappingUtils {
           pluginClass, HubspotOutputFormat.class, SourceInputFormatProvider.class);
     } else if (pluginClass.equals(SalesforceBatchSink.class)) {
       return Plugin.createBatch(
-              pluginClass, SalesforceOutputFormat.class, SalesforceInputFormatProvider.class);
-    }else if (pluginClass.equals(ServiceNowSource.class)) {
+          pluginClass, SalesforceOutputFormat.class, SalesforceInputFormatProvider.class);
+    } else if (pluginClass.equals(ServiceNowSource.class)) {
       return Plugin.createBatch(
           pluginClass, ServiceNowInputFormat.class, SourceInputFormatProvider.class);
     } else if (pluginClass.equals(HubspotStreamingSource.class)) {
@@ -134,7 +134,7 @@ public class MappingUtils {
           checkArgumentNotNull(json, "Can not get JSON from Hubspot input string");
           Object id = json.get(HUBSPOT_ID_FIELD);
           checkArgumentNotNull(id, "Can not get ID from Hubspot input string");
-          return ((Integer) id).longValue();
+          return ((Double) id).longValue();
         } catch (Exception e) {
           LOG.error("Can not get offset from json", e);
         }
